Fix second (and later) runs of Temporal-Color Code
Noticed by Olivier Bardot in bug 606, the *second* run of the temporal
color coder macro will output a list of all LUTs to the Log window with
ImageJ >= 1.47n.

The problem is that we had to work around ImageJ < 1.47n always
returning null instead of the value of the last evaluated statement,
but ImageJ >= 1.47n *does* return said value, and even worse: if it
is not assigned anywhere, the value is printed to the Log window.

To keep the temporal color coder working both with old and new ImageJ
1.x versions (because some plugins do not work with newer ImageJ 1.x
any more), let's put a work-around in place.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>